Added a gltf_uriType tag to each URI indicating which uri type is expected.
I used strings from the data uris to identify the type of the contents.
EX: "data:text/plain;base64, -> text.  We can then use this while
parsing the schema to select an appropriate data type to attempt to
parse the contents of the URI as. (C# EX: byte[] for application, string
for text, and Bitmap for image)