Fix data URI detection to be case-insensitive.
Several changes here, but the functional change is updating the data URI regex to be case-insensitive.

I've also cleaned up the resolvePathIfNeeded function to not rely on a separate, private function, not require a match (which is more expensive than a test), and added some commenting and considerations for future improvements.