Tighten the rules around bufferView.byteStride
The intention of the spec has always been to only use byteStride for
buffer views with vertex data. However, the current language was lax in
a few places, which led to logical contradictions:

- When two accessors with non-vertex data used the same bufferView, it
could be interpreted as requiring byteStride on that view
- However, byteStride was only explicitly allowed for vertex attribute
data, making it seem as if it's not valid on non-vertex views.

glTF-Validator flags bufferViews that have a byteStride but store
non-vertex data as erroneous. Based on this and the above, this change
tightens the language to explicitly only allow byteStride on vertex data
views, and changes all references to it to say the same.