Clarify restrictions on the number of indices
glTF specification is ambiguous as to what the expected behavior is when
a 2-index triangle primitive is specified.

As the spec never defined a valid treatment and this causes continuous
issues with various viewers and processors, this change updates the
spec, making the rules stricter. It is expected that existing assets
largely already conform to these rules as these just spell out the
"logical" requirements of a given mode: it must result in a non-empty
primitive and must not specify any "unused" indices.