Modify "extras" to use same convention as elsewhere in glTF.
I'm not a fan of this, as I'd rather see all "extras" everywhere
get tightened to be an object, rather than the lone object get
loosened to be "any".  But the former is too big a breaking change
to be considered for glTF 2.0, and conformance will help the
glTF Validator enforce uniform behavior, so here we are.