Fixes to khr_character_expression_mapping
Functionally, one-to-many compositions like this are describing how to achieve a particular destination expression, and the way it’s written right now could technically achieve that; but is convoluted. It should be reversed:

<target expression> : { <source expression, weight>, <source expression, weight> }

It also doesn’t make sense to require them to add up to 1.0, as blendshapes don’t really behave that way (and if anything, the weights should reflect how much they contribute to the expression and the max value they should be).

I've made changes to make it clearer, and fixed up the schema (gave it the gltf. prepend and added proper detail as it was really a placeholder before).

Following up with fixing up the schemas elsewhere as well.